requirements.txt
Flask>=3.0,<4
Jinja2>=3.1,<4
Markdown>=3.7,<4
python-frontmatter>=1.1,<2
PyYAML>=6,<7
